Revert "Merge pull request #11 from seohyun1001/sdj"
This reverts commit 21b1e9c8dd44757a121071e4e2662dc937b6d018, reversing
changes made to 6a6ba6b32d0c27d64d93be0b29fe336c95a6f29c.

+++ project_DiB/src/main/java/org/zerock/project_dib/accommodation/dto/AccommodationDTO.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @Builder
 public class AccommodationDTO {
 
-    private Long ano;
+    private int ano;
     private String acc_name;
     private String acc_exp;
     private String acc_loc;
